chore(project): add basic directory layout
run-env: is not fully determined yet, at this point, we only assume:
- python
- django
- postgres

dev-env:
- we won't be using the common approach of project separation into django apps. [1]
- we won't be using PEP-8 code style rules [2]

[1] We think `django.apps` abstraction is mostly for django-reusable-code. For
project-specific-code we use a single app, which name corresponds to the project.

We also use a non common pattern in django of using a separate file for each
entities, in our experience this leads to clear code and reduce edit conflicts.

[2] We use `yapf` with an opinionated set of rules that (in our humble opinion)
produce beautiful python code.

refs #1